Ensure the API response from Riot is not empty before deleting the diff

+++ backend/src/main/java/com/drumonii/loltrollbuild/batch/champions/ChampionsRetrievalItemReader.java
@@ -35,8 +35,10 @@ public class ChampionsRetrievalItemReader extends AbstractItemStreamItemReader<C
 
 	@Override
 	public void open(ExecutionContext executionContext) throws ItemStreamException {
-		List<Champion> deletedChampions = ListUtils.subtract(championsRepository.findAll(), champions);
-		championsRepository.deleteAll(deletedChampions);
+		if (!champions.isEmpty()) {
+			List<Champion> deletedChampions = ListUtils.subtract(championsRepository.findAll(), champions);
+			championsRepository.deleteAll(deletedChampions);
+		}
 	}
 
 }
